config: allow types other than string in get_option()/set_option()
There is no reason to limit options to strings. YAML supports scalars,
lists and dictionaries. As there is no way to write the options set
during runtime back to YAML, so there is no reason to limit the types for
the set_option() method either.

Lift this limitation, so YAML-supported data types can be retrieved and
all data types can be set.

Note that this changes previous behavior, because now options are not
converted to strings anymore.

Signed-off-by: Bastian Krause <bst@pengutronix.de>