remote/exporter: pass proxy information to ResourceEntry
ResourceEntry objects do not have proxy attributes (contrary to
ResourceExports). Resources such as NetworkPowerPorts or ModbusTCPCoils
are ResourceEntry objects, but need this information in case its exporter
is run in isolated mode (-i, --isolated). In this case a SSH port
forwarding is needed to access the device via the exporter.

Allow that by passing the proxy information as part of the ResourceEntry's
data.

Signed-off-by: Bastian Krause <bst@pengutronix.de>