agents/sysfsgpio: Do not clear state in init
Always writing the 'direction' of a GPIO to 'out' leads to the value being reset.
This is not a problem when values are only set.
But: Getting a value will always return 'low'.

This fix checks if the GPIO is already set to 'out' and skips the
re-initialisation in this case.
This way the value of the GPIO is kept and can be read.

Signed-off-by: Chris Fiege <cfi@pengutronix.de>