remote/client: fix forwarding of remote ports
'labgrid-client forward --remote' accidentally uses the variable
'localport', which is either not set at all (leading to a
UnboundLocalError) if no --local/-L is set or is set to the LOCAL part
of --local/-L, which is wrong. The LOCAL part of --remote/-R should be
used instead.

Fix that by using the correct variable 'local'.

Note that the informational print used the correct variable already.

Fixes: 51f3fe5d ("client: Add port forwarding")
Signed-off-by: Bastian Krause <bst@pengutronix.de>