ubootdriver: add configurable sleep before sending interrupt chars
Some U-Boots are configured not to print the "press any key to stop
autoboot" message, but still allow interrupting the boot flow with a key
press. In those cases, another string must be matched (maybe even the
U-Boot version string usually intended to be matched by
boot_expression), which might be printed by U-Boot at a time when the
interrupt prompt is not active yet. Add a configurable sleep interval
after matching the pattern before sending the interrupt to catch those
use cases.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>