doc: configuration: fix docs of sigrok resources
Use the actual resource documented in the corresponding section in the
examples.
Fix the name of the parameter for the channel mapping (channel ->
channels).
Drop the example and extended docs for NetworkSigrokUSBDevice, which are
incorrect. The exporter will take care of that. This should now be
consistent with the other network resources.

Signed-off-by: Bastian Krause <bst@pengutronix.de>