util/managedfile: do not check file creation time in NFS check
Since coreutils 8.31, the file creation time is supported:

  "stat now prints file creation time when supported by the file system,
  on GNU Linux systems with glibc >= 2.28 and kernel >= 4.11." [1]

labgrid checks the file creation (birth) time locally and remotely.
Prior to this change in coreutils' stat, this timestamp was always "0".
Since this change, the actual creation timestamp is displayed on the
local filesystem. Via NFS however, the creation timestamp is not
transferred.

This means that on recent kernels and coreutils versions, stat's output
locally and remotely differ, although the file is actually the same.
This results in an unneeded transfer of the file.

In order to fix this, drop the creation (birth) time from the stat call
in labgrid altogether.

[1] https://lists.gnu.org/archive/html/info-gnu/2019-03/msg00003.html

Signed-off-by: Bastian Krause <bst@pengutronix.de>