remote: add -f,--from for release command
Adds coordinator and client API to release a place, but *only* if it is
acquired by a specific user in an atomic manner.

The intention is that this can be used by cleanup procedures to ensure
that there is not a race condition between figuring out which user has a
resource locked, and actually kicking that user. For example, some CI
systems may leave a place locked if they don't get a chance to cleanup
properly. A cleanup procedure might cross reference the user that has a
place locked with a list of active CI jobs, and if a user is found
that doesn't match any running CI job, the place it has acquired is
released. This presents a race however if this procedure is occuring in
multiple places simultaneously, as it's possible one of them releases
the place, then it gets acquired by an active CI job, then another
cleanup procedure releases it again.

This API prevents the above case because the cleanup procedure can
ensure that the place will *only* be released if it was acquired by the
specific user that it thinks should have it.

Signed-off-by: Joshua Watt <Joshua.Watt@garmin.com>
Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>