Update XMODEM command handling in ShellDriver v2
 - Use return code when running 'which' command in ShellDriver
 - Add 'sz' as a valid option for XMODEM send
 - Reorder tests to find correct version
 - Fixed typos

ShellDriver currently checks for XMODEM by running 'which' on the
desired commands, but uses the stdout to determine success which
doesn't work when 'which' returns "<command> not found" or similar.
This changes the tests to rely on the return code, which is 0 if the
command is located and non-zero for all other cases.

Some systems (Arch) do not provide lsz, but do provide sz to do the
same job. I don't have a BusyBox setup to test, but preserved the rx
command as a final fallback receive command.

https://linux.die.net/man/1/which
https://archlinux.org/packages/community/x86_64/lrzsz/

Signed-off-by: Martin Rowe <martin.p.rowe@gmail.com>
Co-authored-by: Bastian Krause <16608704+Bastian-Krause@users.noreply.github.com>