util/yaml: add tuple constructor to custom loader
This is required for the following commit: switching to the SafeLoader
will not work as is since previous versions of the coordinator will dump
tuples with their `python/tuple` tag. While switching to the SafeDumper
at the same time does solve this problem for any future deployments, it
could result in a breakage if anyone were to upgrade to a later version
of the coordinator without this commit.

Signed-off-by: Ben Brown <ben.brown@codethink.co.uk>