remote/client: no block in async console, loop even on unavailable resource
`labgrid-client console --loop` should keep trying to connect to the
console even if the serial port is currently unvailable.
The console method is declared async because it needs to handle
crossbar communication with the coordinator concurrently, see [1] (e.g.
place kicks, resouces becoming available).

Until now it only tries again if microcom comes back with a return code
!= 0. But in case the resource is unavailable, labgrid bails out early
because target.get_resource() raises a NoResourceFoundError.

Calling target.get_resource() waits until the resources are available by
default. This is done in a blocking poll()/time.sleep() loop until a
timeout is reached allowing no crossbar communication in the meantime.
This again means changes in resource availability we wait for cannot
arrive. So there is no point in waiting for them at all.

Fix that by awaiting the NetworkSerialPort resource while allowing other
tasks to run when --loop is given. Afterwards "wait" for the
NetworkSerialPort for 0 seconds. This either succeeds or a
NoResourceFoundError is raised, both happens instantly.

[1] 19757a18 ("remote/client: close console connection on place release")

Signed-off-by: Bastian Krause <bst@pengutronix.de>