remote/client: update resources before checking NetworkSerialPort availability
PR #935 added async waiting for the NetworkSerialPort. This makes
`labgrid-client console --loop` work as long as the resource changes
from being unavailable to being available. It does not work vice versa,
however: the resources are not being updated before being checked for
availability. The code inside the while loop does not run in this case.
This makes the subsequent `target.await_resources()` raise a
NoResourceFoundError which is not handled anywhere causing an error
message and exit of labgrid-client despite being called with ``--loop``.

Fix that by updating the resources before checking the resource's
availability. `labgrid-client console --loop` will now loop even if the
NetworkSerialPort changes from being available to unavailable.

Fixes: 2230536 ("remote/client: no block in async console, loop even on unavailable resource")
Signed-off-by: Bastian Krause <bst@pengutronix.de>