driver/powerdriver: allow proxying of dynamic port backends with URL in host
Some power backends (e.g. `rest`, `simplerest`) misuse the resource's
host parameter and store a full URL in it. As these backends cannot
specify a `PORT` since it can be a flexible part of the URL specified in
the exporter, such resources cannot be proxied currently.

Now add a method to the NetworkPowerDriver trying to handle the host
parameter as a URL and extract the port from it. Use this port to
request the proxy forward for. Finally construct a proxied variant of
the host parameter to pass to `on()`/`off()`/`get()`.

This allows to use power backends such as `rest` and `simplerest` to be
proxied.

Signed-off-by: Bastian Krause <bst@pengutronix.de>