driver/power/siglent: use backend function signatures expected by NetworkPowerDriver
When using the siglent power backend with the NetworkPowerDriver, this
error occurs:

  self = NetworkPowerDriver(target=Target(name='Test', env=None), name='power', state=<BindingState.active: 2>, delay=2.0)

      @Driver.check_active
      @step()
      def off(self):
  >       self.backend.power_set(self._host, self._port, self.port.index, False)
  E       TypeError: power_set() takes 3 positional arguments but 4 were given

  labgrid/driver/powerdriver.py:219: TypeError

The siglent backend does not follow the function signatures expected by
the NetworkPowerDriver, because it's missing the port parameter. Fix that
by adding the port parameter and making sure that it's None, just like
other backends do (apc, rest, sentry, simplerest, tplink).

Fixes: da0a81d9 ("driver/power: add support for Siglent SPD3000X series power supplies")
Signed-off-by: Bastian Krause <bst@pengutronix.de>