driver/ubootdriver: deprecate boot_expression attribute
Checking the U-Boot output for a expression is problematic for cases
where the driver is activated while U-Boot's auto boot is already
interrupted. This is the case if a strategy wants to force() a U-Boot
state.

A similar check happened to be in the BareboxDriver, but it was dropped
in a refactoring [1].

As a first step in harmonizing BareboxDriver and UBootDriver, deprecate
the boot_expression attribute and do not check for it in
_await_prompt() anymore.

[1] 016f605c ("bareboxdriver: refactor login expect loop and add password support")

Signed-off-by: Bastian Krause <bst@pengutronix.de>