crossbar: separate coordinator guest module arguments
Python allows running a module via 'python -mmodule' (single argument).
To be able to record the coordinator's coverage in a future commit,
'coverage run' is used instead of 'python'. The coverage tool is a
little pickier than python and requires '-m' and the actual module name
to be separate arguments.

Since it will not hurt to have the arguments separated generally, split
them.

Signed-off-by: Bastian Krause <bst@pengutronix.de>