driver/shelldriver: handle bracketed-paste mode
bash >= 5.1 and readline >= 8.1 enable bracketed-paste mode by default,
allowing the terminal emulator to tell a program whether input was typed
or pasted. To achieve this, "\e[?2004h" is inserted when user input is
expected enabling paste detection. "\e[?2004l" is inserted to disable
it.

To handle the escape sequences, match the enable sequence between marker
and prompt. The disable sequence is already stripped by the regex
removing all VT100 codes.

Signed-off-by: Bastian Krause <bst@pengutronix.de>