dynamicnetworkserviceprotocol: Add Protocol
Adds a protocol for creating NetworkService resources at runtime,
typically useful for when some part of the resource can't be known a
priori. For example, QEMU can forward network ports to the guest when
using SLiRP (user) network, but the port assignment can't be know
beforehand. Allowing the SSHDriver to bind to a dynamically generated
NetworkResource allows the forwarding to work seamlessly.

Signed-off-by: Joshua Watt <Joshua.Watt@garmin.com>