labgrid: simplerest - use http PUT command to adapt to powerrelay server
We are using put for setting power. Unsure if anyone has ever used this
for setting power using get; Does it make sense to perform a get to _set_
power ?
If using python3-powerrelay from meta-labgrid, that part works with put
and not get. If anyone are actually using the simplerest to set power,
using get, please say so :)

Signed-off-by: Christian Bräuner Sørensen <christian.brauner.sorensen@prevas.dk>