remote/client: relax SerialDriver requirement for -s/--state
When using -s/--state labgrid deactivates the SerialDriver after
transitioning to the Strategy state to allow exclusive access to the
serial port for microcom. Strategies and environment configs do not
necessarily contain an (active) SerialDriver, e.g. in case only a
SSHDriver is used or a state is requested with no SerialDriver active
yet. Or an ExternalConsoleDriver might be in use instead.

To relax this requirement, try to get an active driver implementing the
ConsoleProtocol. If no such driver is found, continue.

Signed-off-by: Bastian Krause <bst@pengutronix.de>