resource/common: add wrap_command() methods to (Network)Resource
Until now drivers simply prepend the command_prefix for commands that
should work on local and network resources. For local resources, it is
empty, for network resources, it contains
["ssh", .., "<exporter>", "--"].

In some cases this is not enough, because the argument separation is
lost when the command is run via SSH. This means we need shell escaping,
but only for network resources.

So add a wrap_command() method to Resource and NetworkResource. For
Resources, it simply returns the passed command. For NetworkResources,
command_prefix is prepended and each command component is shell-escaped.

Signed-off-by: Bastian Krause <bst@pengutronix.de>