driver/openocddriver: use resource's wrap_command() instead of command_prefix
The OpenOCDDriver runs openocd with `--command` passing an argument
containing spaces: "adapter usb location ..". When run with the SSH
prefix (`command_prefix`), the argument is not received as a single
argument with spaces, but rather as multiple arguments leading
ultimatively to:

  Unexpected command line argument: usb

This issue was hidden with additional quoting until [1]. The current
state only works locally (without the SSH prefix).

To fix this issue, switch from simply prepending the command_prefix to
the new wrap_command() method that takes care of prepending and shell
escaping of the command for NetworkResources.

[1] 875e4881 ("driver/openocddriver: fix command quoting and add tests")

Signed-off-by: Bastian Krause <bst@pengutronix.de>