client: add _get_driver_or_new()
The client code uses the same pattern over and over again:
Get the driver if it is already instanciated via an env config. Else
instanciate a new driver, possibly set the binding map and activate it.

Turn this into a new method while allowing the driver to be specified as
a string. This saves us driver imports in the client code.

Signed-off-by: Bastian Krause <bst@pengutronix.de>