contrib/completion: try to resolve bash aliases
When labgrid-client is called via an alias like..

  alias lgc=labgrid-client

..completion can still be enabled via:

  _completion_loader labgrid-client
  complete -F _labgrid_client lgc

In this case ${base_cmd[0]} is "lgc". Since base_cmd is called for
completing places/resources/matches via the complete sub command, this
does not work, because "lgc" is undefined in the script's scope.

Allow this case by resolving aliases via $BASH_ALIASES.

Signed-off-by: Bastian Krause <bst@pengutronix.de>