remote/exporter: stop event loop when register calls fail
In case an exception is raised during procedure registration, the
default exception handler logs it and the event loop carries on. That
means when the exporter tries to join and it fails to register, the
exporter process simply hangs.

Move the register calls inside the try..except block which already
handles printing the exception and stopping the loop. This way the
process stops and it can be restarted (e.g. by systemd).

Signed-off-by: Bastian Krause <bst@pengutronix.de>