client: console support for --listenonly option with microcom
Expose microcom --listenonly feature via labgrid-client console.
This enables a user to monitor a device serial output
without the risk of client modifying it.
This is useful for multiple concurrent console invocations of same port.

Future amendments to this listenonly feature could enable the usecase
of monitoring a console of a place locked by a different user.

Signed-off-by: Richard Young <code@richyoung.ca>