ssh: Use processwrapper for rsync
Switches to using the processwrapper to run the rsync command in
put_file. This logs the output to the Python logging instead of stdout,
which means it can be better filtered as to where it goes. In
particular, it prevents writing output to stdout (unless desired) when
capsys is disabled.

Signed-off-by: Joshua Watt <Joshua.Watt@garmin.com>