remote/client: allow specifying dfu resources by name
There might be multiple instances of dfu resources in a single place
(either specified via resources with names in the configuration or
via named matches: `labgrid-client add-named-match`).

This allows addressing them via -n/--name. If a configuration is given,
the drivers specified there will be used if they have the same name as
the resource. Otherwise new driver instances will be created.

Signed-off-by: Bastian Krause <bst@pengutronix.de>