logging: Add custom "CONSOLE" log level
The console logging can generate a lot of output at the default level of
INFO, when it may be that all users care about is the step logging. To
deal with this, add a new custom log level called "CONSOLE" that can be
used. This level is below INFO, but above DEBUG.

When invoking pytest on the command line, the verbose flag "-vv", will
show steps (a synonym for "--log-cli-level=INFO"), "-vvv" will show
console input and output (a synonym for "--log-cli-level=CONSOLE") and
"-vvvv" will show almost everything (a synonym for
"--log-cli-level=DEBUG")

Signed-off-by: Joshua Watt <Joshua.Watt@garmin.com>