config: return passed in tool by default
The tools in the tools key can either be a single string with the binary
name or a complete path to a binary. This works for every tool expect
for qemu since drivers can provide a fallback to the default tool.
However qemu currently requires the user to supply the full binary path,
which is unintuitive.

Instead return the originally requested tool-name if no specialized path
is found. And return the simple unexpanded path if a local tool with the
same name can't be found.

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>
(cherry picked from commit 552a42c9f47e325e57e5abf4d368ca17417612ef)
Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>