drop support for Python 3.7
Python 3.7 will be EOL on 2023-06-27 [1]. This is a preparation for
determining labgrid's version. "Use of pkg_resources is deprecated in
favor of [...] importlib.metadata [...]" [2]. See [3].

There is also a backport of this functionality [4], but since the EOL is
not too far away, dropping support for Python 3.7 seems to be the
easiest way.

[1] https://devguide.python.org/versions/
[2] https://setuptools.pypa.io/en/latest/pkg_resources.html
[3] https://docs.python.org/3/library/importlib.metadata.html#module-importlib.metadata
[4] https://pypi.org/project/importlib-metadata/

Signed-off-by: Bastian Krause <bst@pengutronix.de>