Add sshfs exporter class
Adds a class that allows tests to export a local directory to an
exporter via sshfs.

In order to prevent the exporter from being able to access any file that
the user has permission for on the file system (which would allow it to
read SSH keys for example), the SFTP server is run in an isolated
container by default using podman, although this can be overridden to
use docker or directly execute the SFTP server if the exporters are
trusted.

Signed-off-by: Joshua Watt <Joshua.Watt@garmin.com>