github: reusable-unit-tests: check RST syntax of README.rst
Currently only PyPi checks the RST syntax of README.rst because it is
not part of the documentation. This is only noticeable when the version
is pushed to the PyPi test instance, which only happens on master. That
means we notice RST warnings and errors only after a PR is merged to
master.

Change that by checking README.rst explicitly with rstcheck in the unit
test workflow. Install rstcheck just in time to prevent affecting
labgrid's dependencies during install/pytest time.

The `--ignore-languages=bash` argument is needed because the README uses
invalid bash commands such as:

  venv $ pip install .

Signed-off-by: Bastian Krause <bst@pengutronix.de>