drivers, resource: implement get_bound_resources()
Implement a new get_bound_resources() function which returns the
underlying resources a driver is bound to. This can also be called on
Resources, but they will just return themselves in a set.

Add a new keyword argument to get_driver/get_active_driver(), which
allows the user to retrieve a driver using a specific resource it's
bound to.

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>