qemudriver: export get_qemu_base_args method
For debugging, it can be useful to start an interactive session for the
user to access the DUT console. This is possible with real targets
through labgrid-client -s STATE console, but no equivalent exists yet
for targets using QEMUDriver. Resolving that may be a bigger
undertaking, so for now, let's provide a class method
get_qemu_base_args, that returns the list of arguments sans QMP parts.

Users can then initialize labgrid as usual and call the function to
get the command line and start Qemu for interactive use without having
to duplicate the labgrid environment parsing as in [1].

[1]: https://github.com/barebox/barebox/blob/v2023.05.0/test/emulate.pl#L226

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>