qemudriver: support looking up qemu executable in PATH
We currently mandate that a tools key records the full path to the Qemu
binary to use. This is unfortunate as not all distros may install e.g.
qemu-system-arm into the same location. Resolve this by looking up the
qemu_bin key in $PATH if no tools key was found.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>