driver/shelldriver: fix status attribute setting
Since `_check_prompt()` sets the correct `_status` attribute anyway when
a login prompt is detected, this only is a problem for cases without
login (either because no login is required or it happened before).
`get_status()` returns 0 instead of 1 then until `run()` is executed the
first time.

Fix this by setting the correct attribute.

Signed-off-by: Bastian Krause <bst@pengutronix.de>