driver/bareboxdriver: silence barebox in _await_prompt() and unsilence on boot()
Log messages such as..

  eth0: 1000Mbps full duplex link detected

..are logged asynchronously. This confuses the BareboxDriver.

To fix that, set the log level to "emerg" right after a prompt is
detected and recover the previous log level right before booting.

To make the tests run sucessfully, mock the _run() method before driver
activation to act as if `echo $global.loglevel` was run.

Signed-off-by: Bastian Krause <bst@pengutronix.de>