driver/usbstoragedriver: have write_image wait for partition to appear
write_image currently waits for only the block device to appear, even if
the image should be written to a partition. Improve this a bit by
polling for the partition itself to appear with a non-zero size.

While at it, we refactor the code a bit to improve code reuse when we
add the write_files method in the follow-up commit.

Signed-off-by: Stefan Wiehler <stefan.wiehler@missinglinkelectronics.com>
Co-authored-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>