httpdigitaloutput: support generic digital outputs via HTTP
Add support and documentation for a generic digital output driver that
works via HTTP.
The driver allows setting an URL to PUT/POST/PATCH a new output status to
and an URL to GET the current status from.
The message body to send when asserting/deasserting the output can be
specified independently.
The status returned by a GET request can either be matched exactly
against the same body used to set the state or more fuzzily using
regular expressions.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>