client: Allow NetworkService to be manually specified
If a place doesn't have a NetworkService (commonly when the device does
some sort of dynamic IP addressing and tests get the IP address at
runtime), the user needs to manually specify the username, hostname,
port, and password required to SSH into the device. To facilitate this,
add a `--host` and `--password` to the labgrid-client command.

Signed-off-by: Joshua Watt <Joshua.Watt@garmin.com>