util/managedfile: resolve symlinks before comparing NFS paths
ManagedFile has an optimization to detect if the "local" file is
available over NFS at the remote side. In that case, rsync is skipped
and get_remote_path will just return the NFS path.

This is done by comparing stat(1) output of the absolute path on local
and remote host. If the absolute paths contains symlinks, this
comparison may not happen if the remote host doesn't contain the exact
same symlinks. Let's avoid this issue by canonicalizing the local path
first.

Fixes: d42ac82b7a70 ("util/managedfile: heuristically detect NFS to avoid copy")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>