driver/provider: adjust NFSProviderDriver to NFS mounting use cases
During the introduction of the provider drivers, it was assumed all
provider drivers work the same way: by relying on the
`internal`/`external` attributes of their corresponding provider
resource and symlinking the required file (either from
/var/cache/labgrid/$USER/$HASH/file or the given path, in case it's
locally available) to the `internal` path. For the target, the file
is now expected to be available below `external` as returned by
`stage()`.

This, however, does not work for NFS: all NFS servers known to me
don't touch the symlinks, meaning in our case that they point to a
location not exported via NFS. Exporting and mounting these destinations
to the same location is impractical at best or does not work at all.
On top of that, the `internal`/`external` lingo does not fit the NFS
mount use case.

To improve this situation, introduce a custom `stage()` method for the
NFSProviderDriver: assume that /var/cache/labgrid is exported via NFS,
always copy the file to be staged there (even if it would be locally
available) and return an NFSFile object containing the information
required to mount and access the NFS share: host, export and the file
path relative to the export.

Signed-off-by: Bastian Krause <bst@pengutronix.de>